requirements.txt
python-telegram-bot==13.15
python-amazon-paapi
Pillow
requests
schedule
Flask
gunicorn
